Create driver for PIT and some time functions
PIT (Programmable Interval Timer) generate interrupt at regular
interval.
Create a function to set the frequence of this interval and a init
function and add it in kernel_init.

Create time relative functions using pit interface.
A sleep function that return after at least the number of millisecond
given in arguments and a gettime function that return the number of
millisecond since kernel the kernel start (in reality it's since the
interrupts are enabled).

Signed-off-by: xlmod <glafond-@student.42.fr>